web_parser/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.6
loguru==0.7.2
PyYAML==6.0.1